Separate playground state by plugin
